requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic-settings==2.2.1
httpx==0.27.0
jinja2==3.1.4
python-dotenv==1.0.1
pytest==8.2.2
anyio==4.4.0
pypdf